+++ filter/enrolment/lang/et/userdeletefilter_enrolment.php
@@ -1,0 +1,38 @@
+<?php
+// This file is part of Moodle - https://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
+
+/**
+ * Plugin strings are defined here.
+ *
+ * @package     userdeletefilter_enrolment
+ * @category    string
+ * @author      Richard Aljaste <richard.aljaste@ut.ee>
+ * @copyright   2026 University of Tartu <veebitiim@ut.ee>
+ * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+// @codingStandardsIgnoreFile
+
+defined('MOODLE_INTERNAL') || die(); // @codeCoverageIgnore
+
+$string['pluginname'] = 'Kursusele registreerimine';
+$string['privacy:metadata'] = 'See pistikprogramm ei salvesta isikuandmeid.';
+$string['setting_enrolled'] = 'Kasutaja kursusele registreerimise olek';
+$string['setting_enrolled_help'] = 'Valige, kas kasutaja PEAB olema või EI TOHI olla kursusele registreeritud.';
+$string['enrolled'] = 'Peab olema registreeritud vähemalt ühele kursusele';
+$string['enrolled_short'] = 'Registreeritud';
+$string['not_enrolled'] = 'Ei tohi olla registreeritud ühelegi kursusele';
+$string['not_enrolled_short'] = 'Ei ole registreeritud';